requirements.txt
alabaster==0.7.12
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.3.4
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.3
Babel==2.8.0
backcall==0.2.0
bcrypt==3.1.7
bleach==3.1.5
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
cloudpickle==1.5.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-rest-auth==2.1.4
Django==3.0.14
django-allauth==0.44.0
django-countries==7.1
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
flake8==3.8.3
future==0.18.2
gestures==0.2.2
gpg===1.13.1-unknown
helpdev==0.7.1
httplib2==0.14.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
intervaltree==3.0.2
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.5
jupyter-core==4.6.3
kazam==1.4.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
libevdev==0.5
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
nbconvert==5.6.1
nbformat==5.0.7
netifaces==0.10.4
notebook==6.2.0
numpydoc==1.1.0
oauthlib==3.1.0
olefile==0.46
packaging==20.4
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.2
pathtools==0.1.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==2020.11.4
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.5
protobuf==3.6.1
psutil==5.7.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycups==1.9.73
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.6.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.5.3
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.8.0
PyQtWebEngine==5.12.1
pyRFC3339==1.1
pyrsistent==0.16.0
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
python3-openid==3.2.0
pytz==2019.3
pyudev==0.21.0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole==4.7.5
QtPy==1.9.0
redis==3.5.3
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rope==0.17.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.14.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
Sphinx==3.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.1.3
spyder-kernels==1.9.1
sqlparse==0.4.1
systemd-python==234
terminado==0.9.2
testpath==0.4.4
texttable==1.6.2
toml==0.10.1
tornado==6.1
traitlets==4.3.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==1.35
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.0.31
virtualenv-clone==0.5.4
wadllib==1.3.3
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.53.0
wrapt==1.12.1
wurlitzer==2.0.0
xkit==0.0.0
yapf==0.30.0
zipp==1.0.0
